Day05/d05/ex01/views.py
from django.conf import settings
from django.http import HttpRequest, HttpResponse
import psycopg2

def init(request: HttpRequest):
    try:
        return HttpResponse("OK")
    except Exception as e:
        return HttpResponse(e)
